moved sprites function collide_with_walls into sprites.py
